requirements.txt
Django>=5.1.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-allauth>=0.57.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
twilio>=8.0.0
google-auth>=2.20.0
facebook-sdk>=3.1.0
gunicorn>=20.1.0
whitenoise>=6.5.0
python-dotenv>=1.1.0
dj-database-url>=1.0.0
psycopg2-binary>=2.9.0
drf-spectacular>=0.27.0
cryptography>=3.4.0
oauthlib>=3.2.0
requests-oauthlib>=1.3.0
boto3>=1.28.0
django-storages>=1.14.0
awsgi==0.0.5
razorpay>=1.4.0
